backend/requirements.txt
# FastAPI 관련
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7

# AI/ML
numpy==1.24.3
tensorflow==2.15.0
mediapipe==0.10.9
opencv-python==4.8.1.78

# Firebase & 알림
firebase-admin==6.3.0
google-cloud-firestore==2.13.1

# 스케줄링
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 인증 및 보안
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7

# 기타 유용한 패키지
python-multipart==0.0.6
python-dotenv==1.0.0